Move MaterialTableIcons from components/ to util/
