+++ frontend/frontend_app/src/util/MaterialTableIcons.js
